+++ tests/CrEOF/Geo/String/Tests/ParserTest.php
@@ -48,6 +48,17 @@ class ParserTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($expected, $value);
     }
 
+    public function testParserReuse()
+    {
+        $parser = new Parser();
+
+        foreach ($this->dataSourceGood() as $data) {
+            $value = $parser->parse($data[0]);
+
+            $this->assertEquals($data[1], $value);
+        }
+    }
+
     /**
      * @param string $input
      * @param string $exception
@@ -72,6 +83,9 @@ class ParserTest extends \PHPUnit_Framework_TestCase
         return array(
             array('40', 40),
             array('-40', -40),
+            array('1E5', 100000),
+            array('1e5', 100000),
+            array('1e5°', 100000),
             array('40°', 40),
             array('-40°', -40),
             array('40° N', 40),